+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.5.0-alpha.5 — Unreleased
+- Add links to open file in path completions.
+- Add image preview for image files in path completions.
+
 ## 0.5.0-alpha.4 — April 4, 2024
 - Change update links on paste to generate minimal edit.
 - Update github slugifier to more closely match github.